renamed singleinstance.h/cpp to multiprocess.h/cpp
